Introduce pipetool
pipetools provides helper functions for modifying an existing pipeline
on the fly. This allows us to apply wrappers and modifications to the
publishing pipeline within filebeat to be applied to
beat.PipelineConnector directly. This reduces the number of
types/interfaces involved and allows us to apply the same modifications
for the new input v2 API without having to introduce new channel types
or wrappers.