Split internal responsiblit of channl into a RunnerFactory
The channel.OutletFactory wraps a beat.Puplisher pipeline in order to
add shutdown propagation to a beat.Client, and apply default
configurations to all inputs generated.

This changes introduces a wrapper for RunnerFactory, that applies common
input settings to all inputs generated. This allows us to wrap not only
internal inputs, but also the RunnerFactory provided by the
input/v2/compat package.